requirements.txt
fastapi
uvicorn
aiofiles
ultralytics
python-multipart
pillow
opencv-python 